+++ task2/bootstrap/app.ts
@@ -1,0 +1,5 @@
+import express from 'express';
+
+export const app = express();
+
+app.use(express.json());
