+++ task2/index.ts
@@ -1,20 +1,10 @@
-import express from 'express';
 import dotenv from 'dotenv';
 
 dotenv.config();
 
-import { sequelize } from './sequelize';
-import { Validator } from './validator';
-import { UsersDatabase } from './users.database';
-import { UsersController } from './users.controller';
-import { userPartialSchema, userRequiredSchema } from './user.schema';
-
-const app = express();
-const port = process.env.PORT ?? 3000;
-
-const controller = new UsersController(new UsersDatabase());
-
-app.use(express.json());
+import { app } from './bootstrap/app';
+import { sequelize } from './bootstrap/sequelize';
+import { UsersModule } from './modules/users';
 
 app.route('/').get((req, res) => {
     res.send(
@@ -22,16 +12,9 @@ app.route('/').get((req, res) => {
     );
 });
 
-app.route(controller.PREFIX)
-    .get(controller.getAll)
-    .post(Validator.validateAsync(userRequiredSchema), controller.createUser);
+UsersModule.init();
 
-app.route(`${controller.PREFIX}/search`).get(controller.getAutoSuggestUsers);
-
-app.route(`${controller.PREFIX}/:id`)
-    .get(controller.getUser)
-    .patch(Validator.validateAsync(userPartialSchema), controller.updateUser)
-    .delete(controller.deleteUser);
+const port = process.env.PORT ?? 5000;
 
 app.listen(port, () => {
     console.log(`App is listening at http://localhost:${port}`);
@@ -39,9 +22,9 @@ app.listen(port, () => {
 
 sequelize
     .authenticate()
-    .then(() =>
-        console.log('Sequelize: connection has been established successfully.')
-    )
-    .catch((error) =>
-        console.error('Sequelize: unable to connect to the database:', error)
-    );
+    .then(() => {
+        console.log('Sequelize: connection has been established successfully.');
+    })
+    .catch((error) => {
+        console.error('Sequelize: unable to connect to the database:', error);
+    });
